Various improvements in performance to address issue #1: * maintain and use min/max distances from vp for left/right nodes * process nodes within query inside set ahead those with query outside * remember distance outside sets and use to eliminate node later * use points furthest from vp as vantage points for children
These ideas are taken from the NearestNeighbor class in GeographicLib.
The one idea not implemented here is to maintain nodes_to_visit as a
priority queue based on the distance of the query point from the node
(this generalizes the second bullet item above).

On the coast/vessels test, the mean number of distance calculations per
query is
* before = 108.9
* this commit = 16.7
* GeographicLib::NearestNeighbor = 14.7

This commit also eliminate using leaf_size.  The implementation was
buggy:
* it did nothing if the set size was bigger than leaf_size
* otherwise it returned the wrong point

I've left it in as an argument since some of the tests set it.